new update

+++ 0x1E-search_algorithms/0-linear.c
@@ -12,14 +12,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	unsigned int i;
+	int i;
 
 	if (array == NULL || size == 0)
 		return (-1);
 
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int)size; i++)
 	{
-		printf("Value checked array[%u] = [%d]\n", i, array[i]);
+		printf("Value checked array[%d] = [%d]\n", i, array[i]);
 
 		if (array[i] == value)
 			return (i);
